Added vector implementation. Not all features are tested

+++ lib/cpcds_impl/cppstring/cppstring_utils.h
@@ -4,10 +4,10 @@
 #include<stdbool.h>
 #define cpcds_print_cppstr(f,s)fprintf(f,s.bytes)
 #define cpcds_stdprint_cppstr(s)printf(s.bytes)
-int strindex_both_bounds(cppstring*s,char c,unsigned int from,unsigned int to)
+int strindex_both_bounds(cppstring*s,char c,size_t from,size_t to)
 {
 	int index=-1;
-	for(unsigned int i=from;i<to;i++)
+	for(size_t i=from;i<to;i++)
 	{
 		if(str_char_at(s,i)==c)
 		{
@@ -30,7 +30,7 @@ bool str_equal_pointer(cppstring*s,cppstring*t)
 	if(s->len==t->len)
 	{
 		bool e=true;
-		for(unsigned int i=0;i<s->len&&e;i++)
+		for(size_t i=0;i<s->len&&e;i++)
 		{
 			e=str_char_at(s,i)==str_char_at(t,i);
 		}
@@ -47,7 +47,7 @@ bool str_equal_values(cppstring s,cppstring t)
 }
 cppstring str_reverse_order(cppstring*s)
 {
-	for(unsigned int i=0;i<s->len>>1;i++)
+	for(size_t i=0;i<s->len>>1;i++)
 	{
 		str_swap_char(s,i,s->len-i-1);
 	}
@@ -55,7 +55,7 @@ cppstring str_reverse_order(cppstring*s)
 }
 cppstring str_to_upper(cppstring s)
 {
-	for(unsigned int i=0;i<s.len;i++)
+	for(size_t i=0;i<s.len;i++)
 	{
 		if(s.bytes[i]>96||s.bytes[i]<=122)
 		{
@@ -66,7 +66,7 @@ cppstring str_to_upper(cppstring s)
 }
 cppstring str_to_lower(cppstring s)
 {
-	for(unsigned int i=0;i<s.len;i++)
+	for(size_t i=0;i<s.len;i++)
 	{
 		if(s.bytes[i]>64||s.bytes[i]<=90)
 		{
@@ -77,7 +77,7 @@ cppstring str_to_lower(cppstring s)
 }
 cppstring str_replace_char(cppstring*s,char old,char new)
 {
-	for(unsigned int i=0;i<s->len;i++)
+	for(size_t i=0;i<s->len;i++)
 	{
 		if(s->bytes[i]==old)
 		{
@@ -96,7 +96,7 @@ char*cstr_value(cppstring s)
 }
 void cstr_fill_array(cppstring*s,char*str)
 {
-	for(unsigned int i=0;i<s->len;i++)
+	for(size_t i=0;i<s->len;i++)
 	{
 		str[i]=str_char_at(s,i);
 	}
